Fita/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
onnxruntime
numpy
pillow
jinja2
python-multipart